concert_app/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
djangorestframework==3.9.0
idna==2.7
psycopg2==2.7.5
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
